requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python
albumentations
scikit-learn
facenet-pytorch
pillow
kaggle